+++ src/components/onboardingBox.js
@@ -5,6 +5,24 @@ import { T, t } from '../i18n';
 
 const OnboardingBox = React.createClass({
 
+    /*
+                    <p className="fbtrex--opt-in">
+                        {t('dataReuseDesc')}
+                        <a href="https://facebook.tracking.exposed/data" target="_blank">documented and transparent</a>
+                        .
+                    </p>
+                    <div className="clickable">
+                        <span className='default-opt-out' id='data-reuse-button'>
+                            {t('dataReuseButton')}
+                            <br/>
+                            {t('dataReuseButton2')}
+                        </span>
+                        <input type="checkbox" id="data-reuse-checkbox" />
+                    </div>
+
+                    <h2 className="fbtrex--next fbtrex--info" id="ccDesc">{t('optInAtLeast')}</h2>
+     */
+
     render () {
         const publicKey = this.props.publicKey;
 
@@ -20,23 +38,16 @@ const OnboardingBox = React.createClass({
                         <a href="https://facebook.tracking.exposed/privacy-statement" target="_blank">privacy statement</a>
                         .
                     </p>
-                    <button className='default-opt-out' id='info-diet-button'>
-                        {t('infoDietButton')}
-                    </button>
 
-                    <p className="fbtrex--opt-in">
-                        {t('dataReuseDesc')}
-                        <a href="https://facebook.tracking.exposed/data" target="_blank">documented and transparent</a>
-                        .
-                    </p>
+                    <div className="clickable" id="info-diet-button">
+                        <span className='asterisk-default' id='info-diet-asterisk'>* </span>
+                        <span>
+                            {t('infoDietButton')}
+                        </span>
+                        <span className="pretendCheckbox" id="info-diet-checkbox">☐ </span>
+                    </div>
 
-                    <button className='default-opt-out' id='data-reuse-button'>
-                        {t('dataReuseButton')}
-                    </button>
-
-                    <h2 className="fbtrex--next" id="cc">{t('closeAndContinue')}→</h2>
-                    <h2 className="fbtrex--next fbtrex--info" id="ccDesc">{t('optInAtLeast')}</h2>
-
+                    <h2 className="continue-default" id="closeContinue">{t('closeAndContinue')}→</h2>
                 </div>
 
                 <div className='fbtrex--onboarding-collapsed fbtrex--hide'>